+++ src/string/index.ts
@@ -1,3 +1,4 @@
 export * from './email.js'
+export * from './maxCharacters.js'
 export * from './reEmail.js'
 export * from './string.js'