+++ src/string/maxCharacters.test.ts
@@ -1,10 +1,10 @@
 /* eslint-disable import/extensions */
+import { string } from '.'
 import { createValidator } from '../createValidator/createValidator'
-import { string } from '../string'
-import { maxCharactersMessage, maxCharactersYobta } from './'
+import { maxCharacters, maxCharactersMessage } from './maxCharacters'
 
 const customMessage = (limit: number): string => `${limit} yobta!`
-const validate = createValidator(string(), maxCharactersYobta(1, customMessage))
+const validate = createValidator(string(), maxCharacters(1, customMessage))
 
 it('accepts exact lenght', () => {
   const result = validate('a')
@@ -22,7 +22,7 @@ it('regects greater lenght', () => {
 })
 
 it('has default error message', () => {
-  const validateDefault = createValidator(string(), maxCharactersYobta(1))
+  const validateDefault = createValidator(string(), maxCharacters(1))
   const attempt = (): any => validateDefault('ab')
   expect(attempt).toThrow(maxCharactersMessage(1))
 })